No More Default Values
It is in the name, I have fixed the defaultValue glitch that has stalled me from completing the form for the last couple of weeks. Had to initialize the state of the forms to set defaults. Setting certain default values to null will erase them from the data sent to the database, which is good.